+++ src/main/java/playground/controller/document/DocumentController.java
@@ -1,0 +1,29 @@
+package playground.controller.document;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import playground.controller.document.request.CreateDocumentRequest;
+import playground.service.document.DocumentService;
+
+import javax.validation.Valid;
+
+@RestController
+@RequestMapping(path = "/api/documents")
+public class DocumentController {
+
+    private final DocumentService documentService;
+
+    public DocumentController(final DocumentService documentService) {
+        this.documentService = documentService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> create(@RequestBody @Valid final CreateDocumentRequest createDocumentRequest) {
+        documentService.save(createDocumentRequest);
+        return new ResponseEntity<>(HttpStatus.CREATED);
+    }
+}
