--- src/main/java/playground/domain/ApprovalState.java
@@ -1,15 +1,0 @@
-package playground.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum ApprovalState {
-
-    DRAFTING("결재중"),
-    APPROVED("승인"),
-    CANCELED("거절");
-
-    private final String text;
-}
