+++ src/main/java/playground/domain/document/ApprovalState.java
@@ -1,0 +1,15 @@
+package playground.domain.document;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ApprovalState {
+
+    DRAFTING("결재중"),
+    APPROVED("승인"),
+    CANCELED("거절");
+
+    private final String text;
+}
